Models for fluids

Newtonian fluid

MaterialModels.NewtonianFluid — Type

NewtonianFluid(::WaterEOS; μ, λ = -2μ/3)

Equation

\[\sigma_{ij} = -p \delta_{ij} + \lambda d_{kk} \delta_{ij} + 2\mu d_{ij},\]

where $p$ is the pressure and $d_{ij}$ is the rate of deformation tensor.

Parameters

μ: dynamic viscosity

λ: second coefficient of viscosity

Methods

MaterialModels.matcalc__pressure__ρ — Method

@matcalc(:pressure, model::NewtonianFluid{EOS}; ρ::Real)

Compute the pressure based on the equation of state EOS.

MaterialModels.matcalc__density__p — Method

@matcalc(:pressure, model::NewtonianFluid{EOS}; p::Real)

Compute the mass density based on the equation of state EOS.

MaterialModels.matcalc__stress__d__ρ — Method

@matcalc(:stress, model::NewtonianFluid; d::SymmetricSecondOrderTensor{3}, ρ::Real)

Compute the stress.