Functions

MaterialModels.initial_material_state — Function

initial_material_state(::AbstractMaterial)

Return the MaterialState that belongs to the given Material and is initialized with zeros.

MaterialModels.get_cache — Function

get_cache(m::AbstractMaterial)

Construct cache object for iteratively solving non-linear material models.

For material models which require an iterative solution procedure, it is recommended to allocate storage for the iterative solver only once and reuse it for all material points. When multithreading is used, each threads needs its own cache.

Returns nothing for materials that don't need a cache.

MaterialModels.update_cache! — Function

update_cache!(cache::OnceDifferentiable, f)

Update the cache object with the residual function for the current time/load step.

As the residual functions depend i.a. on the strain increment, the function and its jacobian need to be updated for every load step.