LinearElastic

MaterialModels.LinearElastic — Type

LinearElastic(E, ν)

Isotropic linear elasticity.

Arguments

E::Float64: Young's modulus

ν::Float64: Poisson's ratio

MaterialModels.material_response — Method

material_response(m::LinearElastic, ε::SymmetricTensor{2,3})

Return the stress tensor and the stress tangent for the given strain ε such that

\[\boldsymbol{\sigma} = \mathbf{E}^\text{e} : \boldsymbol{\varepsilon} .\]

No MaterialState is needed for the stress computation, thus if a state is handed over to material_response, the same state is returned.