TransverselyIsotropic

MaterialModels.TransverselyIsotropic — Type

TransverselyIsotropic(; E_L, E_T, G_LT, ν_LT, ν_TT)

Transversely isotropic elasticity.

The material direction (the vector normal to the symmetry plane) is specified in TransverselyIsotropicState, which is constructed with initial_material_state(::TransverselyIsotropic, direction::Vec{3}). The default direction is [1.0, 0.0, 0.0].

Arguments

E_L::Float64: LongitudinalYoung's modulus

E_T::Float64: Transverse Young's modulus

G_LT::Float64: In-plane shear modulus

ν_LT::Float64: In-plane Poisson's ratio

ν_TT::Float64: Out-of-plane Poisson's ratio

MaterialModels.material_response — Method

material_response(m::TransverselyIsotropic, ε::SymmetricTensor{2,3}, state::TransverselyIsotropicState)

Return the stress tensor and the stress tangent for the given strain ε such that

\[\boldsymbol{\sigma} = \mathbf{E} : \boldsymbol{\varepsilon}\]

where

\[\mathbf{E} = L_\perp \boldsymbol{I} \otimes \boldsymbol{I} + [L_\parallel - L_\perp][\boldsymbol{I}\otimes\boldsymbol{A} + \boldsymbol{A}\otimes\boldsymbol{I}] + [M_\parallel - 4G_\parallel + 2G_\perp - 2L_\parallel + L_\perp]\boldsymbol{A}\otimes\boldsymbol{A} + 4[G_\parallel-G_\perp]\mathbf{A}\]

\[\mathbf{A} = \frac{1}{4} (\boldsymbol{A} \overline{\otimes} \boldsymbol{I} + \boldsymbol{A} \underline{\otimes} \boldsymbol{I} + \boldsymbol{I} \overline{\otimes} \boldsymbol{A} + \boldsymbol{I} \underline{\otimes} \boldsymbol{A}) \]

\[\boldsymbol{A} = \boldsymbol{a} \otimes \boldsymbol{a}\]

and where $\boldsymbol{a}$ is the vector normal to the plane of symmetry.