XuNeedleman

MaterialModels.XuNeedleman — Type

XuNeedleman(σₘₐₓ, τₘₐₓ, Φₙ, Φₜ, Δₙˢ)

Xu-Needleman traction-separation law.[Xu1993] This is a commonly used cohesive law for brittle fracture, however it is revertible and thus not suited for unloading. In 3D, it is isotropic within the cohesive plane.

Arguments

σₘₐₓ::Float64: cohesive normal strength

τₘₐₓ::Float64: cohesive tangential strength (in-plane strength)

Φₙ::Float64: normal work of separation

Φₜ::Float64: tangential work of separation

Δₙˢ: normal separation after complete shear separation under the condition of zero normal tension

For convenience the following functions can be used to convert parameters:

xu_needleman_Φₙ

xu_needleman_Φₜ

xu_needleman_σₘₐₓ

xu_needleman_τₘₐₓ

MaterialModels.material_response — Method

material_response(m::XuNeedleman, Δ::Tensor{1,dim}) where dim

Return the traction vector and the traction tangent for the given separation jump Δ. The last entry of Δ is interpreted as normal separation, the first entries are interpreted as in-plane separations.

No MaterialState is needed for the stress computation, thus if a state is handed over to material_response, the same state is returned.

MaterialModels.xu_needleman_Φₙ — Method

xu_needleman_Φₙ(σₘₐₓ, δₙ)

Compute the normal work of separation Φₙ for the Xu-Needleman cohesive law based on the cohesive normal strength σₘₐₓ and the characteristic normal separation δₙ.

MaterialModels.xu_needleman_Φₜ — Method

xu_needleman_Φₜ(τₘₐₓ, δₜ)

Compute the tangential work of separation Φₜ for the Xu-Needleman cohesive law based on the cohesive tangential strength τₘₐₓ and the characteristic tangential separation δₜ.

MaterialModels.xu_needleman_σₘₐₓ — Method

xu_needleman_σₘₐₓ(Φₙ, δₙ) = Φₙ / (exp(1.0) * δₙ)

Compute the cohesive normal strength σₘₐₓ for the Xu-Needleman cohesive law based on the normal work of separation Φₙ and the characteristic normal separation δₙ.

MaterialModels.xu_needleman_τₘₐₓ — Method

xu_needleman_τₘₐₓ(Φₜ, δₜ)

Compute the cohesive tangential strength τₘₐₓ for the Xu-Needleman cohesive law based on the tangential work of separation Φₜ and the characteristic tangential separation δₜ.

Xu1993Xu, X. P., & Needleman, A. (1993). Void nucleation by inclusion debonding in a crystal matrix. Modelling and Simulation in Materials Science and Engineering, 1(2), 111–132. https://doi.org/10.1088/0965-0393/1/2/001