Utilities

MaterialModels.matcalc__jaumann2caucy__W__dσ_jaumann__σ — Method

@matcalc(:jaumann2caucy; dσ_jaumann::SymmetricSecondOrderTensor{3}, σ::SymmetricSecondOrderTensor{3}, W::SecondOrderTensor{3})

Convert the Jaumann stress rate to the Caucy stress rate. For example, this function can be used as follows:

dσᴶ = @matcalc(:stress, model; σ, dϵ = symmetric(∇v*dt)) - σ
dσ = @matcalc(:jaumann2caucy; dσ_jaumann = dσᴶ, σ, W = skew(∇v*dt))

Equation

\[\dot{\bm{\sigma}}^\mathrm{J} = \dot{\bm{\sigma}} - \bm{W} \cdot \bm{\sigma} + \bm{\sigma} \cdot \bm{W}\]

MaterialModels.matcalc__soundspeed__E__ν__ρ — Method

@matcalc(:soundspeed; K::Real, G::Real, ρ::Real)

Compute the speed of sound in solids.

Equation

\[c = \sqrt{\frac{E(1-\nu)}{\rho (1+\nu)(1-2\nu)}}\]

Parameters

E: Young's modulus

ν: Poisson's ratio

ρ: density

MaterialModels.matcalc__soundspeed__G__K__ρ — Method

@matcalc(:soundspeed; K::Real, G::Real, ρ::Real)

Compute the speed of sound in solids.

Equation

\[c = \sqrt{\frac{K + \frac{4}{3}G}{\rho}}\]

Parameters

K: bulk modulus

G: shear modulus

ρ: density