TransverselyIsotropic

Missing docstring.

Missing docstring for TransverselyIsotropic. Check Documenter's build log for details.

MaterialModels.material_response — Method

material_response(m::TransverselyIsotropic, ε::SymmetricTensor{2,3}, state::TransverselyIsotropicState)

Return the stress tensor and the stress tangent for the given strain ε such that

\[\boldsymbol{\sigma} = \mathbf{E} : \boldsymbol{\varepsilon}\]

where

\[\mathbf{E} = L_\perp \boldsymbol{I} \otimes \boldsymbol{I} + [L_\parallel - L_\perp][\boldsymbol{I}\otimes\boldsymbol{A} + \boldsymbol{A}\otimes\boldsymbol{I}] + [M_\parallel - 4G_\parallel + 2G_\perp - 2L_\parallel + L_\perp]\boldsymbol{A}\otimes\boldsymbol{A} + 4[G_\parallel-G_\perp]\mathbf{A}\]

\[\mathbf{A} = \frac{1}{4} (\boldsymbol{A} \overline{\otimes} \boldsymbol{I} + \boldsymbol{A} \underline{\otimes} \boldsymbol{I} + \boldsymbol{I} \overline{\otimes} \boldsymbol{A} + \boldsymbol{I} \underline{\otimes} \boldsymbol{A}) \]

\[\boldsymbol{A} = \boldsymbol{a} \otimes \boldsymbol{a}\]

and where $\boldsymbol{a}$ is the vector normal to the plane of symmetry.